+++ src/components/App.js
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import Header from './Header';
 import TextBox from './TextBox';
 import Results from './Results';
 
@@ -14,14 +15,7 @@ function App() {
   return (
     <main className="my-12 sm:my-20">
       <div className="container text-center">
-        <div className="mb-12">
-          <h1 className="mb-1 title-h1 text-orange">
-            Word Limit
-          </h1>
-          <h3>
-            A simple word and character counter. Free to use.
-          </h3>
-        </div>
+        <Header />
         <Results text={text} />
         <TextBox text={text} handleOnChange={handleOnChange} />
       </div>