+++ CHANGELOG.md
@@ -1,10 +1,23 @@
 # Changelog
 
-Release notes for the Word Limit app.
+Release notes for the Count My Words app.
 
 
 
 ## Unreleased
+
+### Added
+- Added SEO title and description.
+
+### Changed
+- Changed the name of the app to *Count My Words*.
+
+### Fixed
+- Fixed layout of counters area on small screens.
+
+
+
+## 0.2.0 - 2022-07-29
 
 ### Added
 - Added the ability to count sentences and paragraphs.